+++ src/whatrecord/shell.py
@@ -1,6 +1,8 @@
 import functools
+import inspect
 import logging
 import multiprocessing as mp
+import os
 import pathlib
 import sys
 import time
@@ -81,9 +83,11 @@ class ShellState(ShellStateBase):
         Current loading context stack (e.g., ``st.cmd`` then ``common_startup.cmd``)
     """
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
+    def __post_init__(self):
+        from .macro import MacroContext
+        self.macro_context = MacroContext()
+        self.shell = None
+        self.handlers = dict(self.find_handlers())
         self._setup_dynamic_handlers()
 
     def _setup_dynamic_handlers(self):
@@ -93,6 +97,51 @@ class ShellState(ShellStateBase):
                 self.handlers[name] = functools.partial(
                     self._generic_motor_handler, name
                 )
+
+    def find_handlers(self):
+        for attr, obj in inspect.getmembers(self):
+            if attr.startswith("handle_") and callable(obj):
+                name = attr.split("_", 1)[1]
+                yield name, obj
+
+    def get_asyn_port_from_record(self, inst: RecordInstance):
+        rec_field = inst.fields.get("INP", inst.fields.get("OUT", None))
+        if rec_field is None:
+            return
+
+        value = rec_field.value.strip()
+        if value.startswith("@asyn"):
+            try:
+                asyn_args = value.split("@asyn")[1].strip(" ()")
+                asyn_port, *_ = asyn_args.split(",")
+                return self.asyn_ports.get(asyn_port.strip(), None)
+            except Exception:
+                logger.debug("Failed to parse asyn string", exc_info=True)
+
+    def get_frozen_context(self):
+        if not self.load_context:
+            return tuple()
+        return tuple(ctx.freeze() for ctx in self.load_context)
+
+    def handle_command(self, command, *args):
+        handler = self.handlers.get(command, None)
+        if handler is not None:
+            return handler(*args)
+        return self.unhandled(command, args)
+        # return f"No handler for handle_{command}"
+
+    def _fix_path(self, filename: str):
+        if os.path.isabs(filename):
+            for from_, to in self.standin_directories.items():
+                if filename.startswith(from_):
+                    _, suffix = filename.split(from_, 1)
+                    return pathlib.Path(to + suffix)
+
+        return self.working_directory / filename
+
+    def unhandled(self, command, args):
+        ...
+        # return f"No handler for handle_{command}"
 
     def _generic_motor_handler(self, name, *args):
         arg_info = []
@@ -426,7 +475,7 @@ def load_startup_scripts(
             return container
 
     loader = functools.partial(load_startup_script, standin_directories)
-    for idx, fn in enumerate(sorted(set(fns))):
+    for idx, fn in enumerate(sorted(set(fns)), 1):
         t0 = time.monotonic()
         if idx == 20:
             break
